XftFontOpenInfo: Use of uninitialised value of size 8 (bug 11200)
This is due to XftFontInfoFill using the binary representation of the
XftFontInfo to generate fi->hash.

With 64-bit pointers there is padding between .hash and .file in struct
_XftFontInfo.  This padding is not initialized, and the hash uses these
bytes.

This will interfere with finding "a matching previously opened font" in
XftFontOpenInfo, and XftFontInfoEqual, which uses memcmp, will have similar
problems.

This fix makes no assumptions about the sizes and alignment of members of
struct _XftFontInfo by using memset.  (It also makes no assumptions about
what FcPatternGet* does to its output parameter when it returns
FcResultNoMatch.)