Skip 'render' pattern elements with invalid type
Fontconfig no longer supports the FcNameRegisterObjectTypes API, which
means that any Xft custom fields cannot be specified in a fontconfig
name as those will always be set to FcTypeUnknown in the pattern.

The only one of those we care about is XFT_RENDER, which was used to
disable the Render extension, something most apps probably don't need
to support in names anyways. Allow the call requesting this value to
return FcResultTypeMismatch and pretend that XFT_RENDER is missing
from the pattern.

Signed-off-by: Keith Packard <keithp@keithp.com>